Reviewed by Hyatt.
        <rdar://problem/5193517> FrameLoader::canCachePage() might walk the entire DOM
        http://bugs.webkit.org/show_bug.cgi?id=13662

        * loader/FrameLoader.cpp:
        (WebCore::FrameLoader::provisionalLoadStarted): Remove some useless and now harmful
        code - there's no need to clear in this case.



Canonical link: https://commits.webkit.org/17867@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@21421 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/loader/FrameLoader.cpp
@@ -1551,16 +1551,9 @@ void FrameLoader::provisionalLoadStarted()
     cancelRedirection(true);
     m_client->provisionalLoadStarted();
 
-    if (canCachePage()) {
-        if (m_client->canCachePage()) {
-            if (!m_currentHistoryItem->cachedPage()) {
-                cachePageToHistoryItem(m_currentHistoryItem.get());
-                purgePageCache();
-            }
-        } else {
-            // Put the document into a null state, so it can be restored correctly. 
-            clear();
-        }
+    if (canCachePage() && m_client->canCachePage() && !m_currentHistoryItem->cachedPage()) {
+        cachePageToHistoryItem(m_currentHistoryItem.get());
+        purgePageCache();
     }
 }
